Crossbuild libusb for android using NDK
Added build.sh script for cross compilation of libusb for android. This script
uses standalone toolchain of Android NDK to generate libusb for android. All the
generated libraries go in the location : crossbuild/ndk-$ARCH/sysroot/usr/lib
where ARCH is either x86 or arm depending upon the architecture one wants to
build on. The default is x86 as the tests are going to be conducted on emulator.
The built library of libusb would be used later by native code on android.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>