Modified Home to call native code
Now, on start of activity, it calls the native code openUsbDevice (to be
implemented in open_device.c file). This function returns 0 for no error and
negative values for error. Accordingly the background of the activity changes to
Green or Red. Logs can be seen on logcat for more information.

According to my experiment, there are two common scenarios.
1. I open my activity directly by opening the application. Here, I would not
have explicit permissions to use the USB device, as I have devised no way to do
so. So, the background should turn RED everytime.
2. I attach the USB device, it asks for permission to use USB device in this
activity. I give it permission. The activity opens. Now the question is, would
everything go as per plan and the FTDI USB device will be opened easily by
libusb - The background turns GREEN (or) I still do not have access or some
other error - The background turns RED.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>