Added open_usb.c to open USB device matching FTDI
Listing all the USB devices and stepping at each of them, seeing if they are
FTDI chips or not ( by matching VID and PID). And if they are, open them. If
there is lack of appropriate permissions, this will fail and -1 would be
returned. Also, mechanism included to map native function open_usb_device to
java function openUsbDevice by using JNI registerNatives method.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>